refact: enhance Button component to support icons and update layout styles; add GitOverview and ImageContainer components with styling; update translations for GitOverview
