+++ src/pages/HomeSection/components/ImageContainer/consts.ts
