+++ src/themes/index.ts
@@ -39,7 +39,7 @@ export const themes = {
     XLARGE: '24px',
     XXLARGE: '32px',
     XXXLARGE: '48px',
-    XXXXLARGE: '84px',
+    XXXXLARGE: '72px',
   },
   FONT_WEIGHT: {
     LIGHT: 300,
